# backend/apiService/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /collection:
        get:
            tags:
                - CollectionService
            description: 列出用户的收藏夹
            operationId: CollectionService_ListCollection
            parameters:
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ListCollectionResponse'
        put:
            tags:
                - CollectionService
            description: 更新收藏夹信息
            operationId: CollectionService_UpdateCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.UpdateCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.UpdateCollectionResponse'
        post:
            tags:
                - CollectionService
            description: 创建收藏夹
            operationId: CollectionService_CreateCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.CreateCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.CreateCollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 删除收藏夹
            operationId: CollectionService_RemoveCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.RemoveCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.RemoveCollectionResponse'
    /collection/media:
        get:
            tags:
                - CollectionService
            description: 列出收藏夹中的媒体
            operationId: CollectionService_ListMedia4Collection
            parameters:
                - name: collectionId
                  in: query
                  description: '@gotags: json:"collectionId,omitempty,string"'
                  schema:
                    type: string
                - name: pagination.page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.size
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.ListMedia4CollectionResponse'
        post:
            tags:
                - CollectionService
            description: 将媒体添加到收藏夹
            operationId: CollectionService_AddMedia2Collection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.AddMedia2CollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.AddMedia2CollectionResponse'
        delete:
            tags:
                - CollectionService
            description: 从收藏夹中移除媒体
            operationId: CollectionService_RemoveMediaFromCollection
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.RemoveMediaFromCollectionRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.RemoveMediaFromCollectionResponse'
...
    /favorite:
        post:
            tags:
                - FavoriteService
            operationId: FavoriteService_AddFavorite
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.AddFavoriteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.AddFavoriteResponse'
        delete:
            tags:
                - FavoriteService
            operationId: FavoriteService_RemoveFavorite
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/apiService.svapi.RemoveFavoriteRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/apiService.svapi.RemoveFavoriteResponse'
...
components:
    schemas:
        apiService.svapi.AddFavoriteRequest:
            type: object
            properties:
                target:
                    type: integer
                    format: enum
                type:
                    type: integer
                    format: enum
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
        apiService.svapi.AddFavoriteResponse:
            type: object
            properties: {}
...
        apiService.svapi.AddMedia2CollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                mediaId:
                    type: string
                    description: '@gotags: json:"mediaId,omitempty,string"'
        apiService.svapi.AddMedia2CollectionResponse:
            type: object
            properties: {}
...
        apiService.svapi.Collection:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                userId:
                    type: string
                    description: '@gotags: json:"userId,omitempty,string"'
                name:
                    type: string
                description:
                    type: string
        apiService.svapi.CollectionMedia:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                mediaId:
                    type: string
                    description: '@gotags: json:"mediaId,omitempty,string"'
                coverUrl:
                    type: string
                title:
                    type: string
                description:
                    type: string
                mediaUrl:
                    type: string
...
        apiService.svapi.CreateCollectionRequest:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
        apiService.svapi.CreateCollectionResponse:
            type: object
            properties:
                collection:
                    $ref: '#/components/schemas/apiService.svapi.Collection'
...
        apiService.svapi.ListCollectionResponse:
            type: object
            properties:
                collections:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.Collection'
                pagination:
                    $ref: '#/components/schemas/apiService.svapi.PaginationResponse'
...
        apiService.svapi.ListMedia4CollectionResponse:
            type: object
            properties:
                medias:
                    type: array
                    items:
                        $ref: '#/components/schemas/apiService.svapi.CollectionMedia'
                pagination:
                    $ref: '#/components/schemas/apiService.svapi.PaginationResponse'
...
        apiService.svapi.PaginationResponse:
            type: object
            properties:
                page:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                count:
                    type: integer
                    format: int32
...
        apiService.svapi.RemoveCollectionRequest:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
        apiService.svapi.RemoveCollectionResponse:
            type: object
            properties: {}
...
        apiService.svapi.RemoveFavoriteRequest:
            type: object
            properties:
                target:
                    type: integer
                    format: enum
                type:
                    type: integer
                    format: enum
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
        apiService.svapi.RemoveFavoriteResponse:
            type: object
            properties: {}
...
        apiService.svapi.RemoveMediaFromCollectionRequest:
            type: object
            properties:
                collectionId:
                    type: string
                    description: '@gotags: json:"collectionId,omitempty,string"'
                mediaId:
                    type: string
                    description: '@gotags: json:"mediaId,omitempty,string"'
        apiService.svapi.RemoveMediaFromCollectionResponse:
            type: object
            properties: {}
...
        apiService.svapi.UpdateCollectionRequest:
            type: object
            properties:
                id:
                    type: string
                    description: '@gotags: json:"id,omitempty,string"'
                name:
                    type: string
                description:
                    type: string
        apiService.svapi.UpdateCollectionResponse:
            type: object
            properties:
                collection:
                    $ref: '#/components/schemas/apiService.svapi.Collection'
...